Job file cleanup
* Moved base job class into its own file
* Added storing of error description to job class
